requirements.txt
streamlit==1.32.0
pandas==2.2.0
pillow==10.2.0
mysql-connector-python==8.3.0
python-dateutil==2.8.2 